✂️  split ILogger interface into ILogger and ILog
