+++ modules/client/src/connect.ts
@@ -2,7 +2,7 @@ import "core-js/stable";
 import "regenerator-runtime/runtime";
 
 import { IMessagingService, MessagingServiceFactory } from "@connext/messaging";
-import { CF_PATH, CREATE_CHANNEL_EVENT, ILogger, StateSchemaVersion } from "@connext/types";
+import { CF_PATH, CREATE_CHANNEL_EVENT, ILog, ILogger, StateSchemaVersion } from "@connext/types";
 import { Contract, providers } from "ethers";
 import { AddressZero } from "ethers/constants";
 import { fromExtendedKey, fromMnemonic } from "ethers/utils/hdnode";
